==> 03_elected_candidates.ipynb <==
import requests # Downloading webpages
from bs4 import BeautifulSoup # Extracting data from html files
import json # Parsing JSON
from pathlib import Path # Working with files
import pandas as pd # Dataframes

url = 'https://www.dr.dk/nyheder/politik/folketingsvalg/valgte' # website with links to all elected candidates
req = requests.get(url) # download page
soup = BeautifulSoup(req.text) # using .text extracts html code from webpage

str_json = soup.find('script', id='__NEXT_DATA__').text # get json part
elected_file = Path('data/elected_file.json') # path for file
elected_file.write_text(str_json) # save json

elected_list = []

with open('data/elected_file.json', "r") as f: # open json file
        elected_json = json.load(f)        

for i in range(175):  # for all elected candidates, save:
    elected_data = {
    "ID" : elected_json["props"]["pageProps"]["electedCandidates"][i]["urlkey"], # ID
    "votes" : elected_json["props"]["pageProps"]["electedCandidates"][i]["v"]} # number of votes

    elected_list.append(elected_data) # append data for one candidate to the rest

    
elected = pd.DataFrame(elected_list) # make dict into dataframe
elected.to_csv(Path('data/kandidat_elected.csv')) # save dataframe
